fix(ci): python-coverage-shards — split test files over the 1,500-line maintainability limit
This PR's added tests pushed three test modules past the
MAX_FIRST_PARTY_FILE_LINES=1500 guardrail asserted by
test_first_party_code_files_stay_under_line_limit, failing
python-coverage-shards (8). Split each oversized module under the limit
without weakening the gate or dropping any test:

- runtime/test_planning_parts/test_planning_part_001.py (1543): move the
  trailing classify_conformance_stall block into the sibling
  test_planning_part_002.py, which already hosts the stall suite and its
  _stall_policy/_iter_record doubles; drop the now-dead helpers and
  unused imports left behind.
- control/test_planning_ops_branch_edges.py (1503): extract the
  self-contained _build_conformance_stall_failure tests into a new
  test_planning_ops_stall_failure.py.
- runtime/test_validation_worktree.py (1609): extract the real-git
  gitlink/submodule tests into a new test_validation_worktree_gitlink.py.

Test name sets are preserved exactly (77/29/48 across the three groups)
and every first-party file is back under 1,500 lines.
