+++ tests/unit/runtime/test_planning_parts/test_planning_part_002.py
@@ -6,6 +6,10 @@ from pathlib import Path
 
 import pytest
 
+from awf.adapters.base import AgentRunError
+from awf.common.commands import CommandResult
+from awf.db.enums import AgentRuntime
+from awf.runtime import planning as planning_mod
 from awf.runtime.planning import (
     AGENT_STALLED_IN_CONFORMANCE,
     MAX_CONFORMANCE_TEXT_CHARS,
@@ -13,6 +17,7 @@ from awf.runtime.planning import (
     ConformanceStallEvidence,
     ConformanceStallKind,
     ConformanceStallPolicy,
+    _evidence_strings,
     build_conformance_stall_failure_evidence,
     build_conformance_stall_recovery_prompt,
     classify_conformance_stall,
@@ -409,6 +414,364 @@ def test_classify_conformance_stall_breaks_no_output_streak_on_report_file_progr
         policy=_stall_policy(no_output_seconds=600),
         plan_path=Path("docs/awf-plans/ws_report_progress.md"),
         report_path=Path("docs/awf-plans/ws_report_progress.conformance.json"),
+        latest_error=None,
+    )
+
+    assert evidence is None
+
+
+@pytest.mark.unit
+def test_classify_conformance_stall_returns_no_output_for_idle_timeout() -> None:
+    history = [
+        _iter_record(
+            iteration=0,
+            elapsed_seconds=120.0,
+            report_digest="abc",
+            worktree_changed=True,
+            stdout="some output",
+        ),
+        _iter_record(
+            iteration=1,
+            elapsed_seconds=620.0,
+            report_digest=None,
+            worktree_changed=False,
+            stdout="" * 0,
+            stderr="",
+            error_reason_code="AGENT_IDLE_TIMEOUT",
+        ),
+    ]
+    error = AgentRunError(
+        agent=AgentRuntime.codex,
+        result=CommandResult(returncode=124, stdout="", stderr="idle timeout exceeded"),
+        reason_code="AGENT_IDLE_TIMEOUT",
+    )
+
+    evidence = classify_conformance_stall(
+        history=history,
+        policy=_stall_policy(),
+        plan_path=Path("docs/awf-plans/ws_no_output.md"),
+        report_path=Path("docs/awf-plans/ws_no_output.conformance.json"),
+        latest_error=error,
+    )
+
+    assert evidence is not None
+    assert evidence.kind == ConformanceStallKind.no_output
+    assert evidence.iteration_index == 1
+    assert evidence.elapsed_seconds == pytest.approx(620.0 + 120.0)
+    assert evidence.no_output_seconds == pytest.approx(620.0)
+    assert evidence.repeated_output_count == 0
+    assert evidence.last_report_digest is None
+    assert evidence.plan_path == "docs/awf-plans/ws_no_output.md"
+    assert evidence.report_path == "docs/awf-plans/ws_no_output.conformance.json"
+    assert "idle timeout" in evidence.last_output_excerpt
+
+
+@pytest.mark.unit
+def test_classify_conformance_stall_returns_none_without_history() -> None:
+    assert (
+        classify_conformance_stall(
+            history=[],
+            policy=_stall_policy(),
+            plan_path=Path("docs/awf-plans/ws_empty.md"),
+            report_path=Path("docs/awf-plans/ws_empty.conformance.json"),
+            latest_error=None,
+        )
+        is None
+    )
+
+
+@pytest.mark.unit
+def test_classify_conformance_stall_uses_error_stdout_when_stderr_is_empty() -> None:
+    history = [
+        _iter_record(
+            iteration=0,
+            elapsed_seconds=600.0,
+            report_digest=None,
+            worktree_changed=False,
+            stdout="",
+            error_reason_code="AGENT_TIMEOUT",
+        )
+    ]
+    error = AgentRunError(
+        agent=AgentRuntime.codex,
+        result=CommandResult(returncode=124, stdout="stdout-only failure", stderr=""),
+        reason_code="AGENT_TIMEOUT",
+    )
+
+    evidence = classify_conformance_stall(
+        history=history,
+        policy=_stall_policy(over_duration_seconds=300),
+        plan_path=Path("docs/awf-plans/ws_stdout.md"),
+        report_path=Path("docs/awf-plans/ws_stdout.conformance.json"),
+        latest_error=error,
+    )
+
+    assert evidence is not None
+    assert evidence.kind == ConformanceStallKind.over_duration
+    assert "stdout-only failure" in evidence.last_output_excerpt
+
+
+@pytest.mark.unit
+def test_classify_conformance_stall_redacts_secrets_in_last_output_excerpt() -> None:
+    # Stall evidence is persisted as durable failure details and surfaced into
+    # recovery prompts; raw agent stderr/stdout can include provider tokens or
+    # URL credentials. Verify the excerpt is scrubbed before persistence.
+    leaked_token = "ghp_AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA"
+    leaked_stderr = f"fatal: could not read Username for 'https://github.com/': {leaked_token}"
+    history = [
+        _iter_record(
+            iteration=0,
+            elapsed_seconds=30.0,
+            report_digest=None,
+            worktree_changed=False,
+            stderr=leaked_stderr,
+            error_reason_code="AGENT_IDLE_TIMEOUT",
+        ),
+    ]
+    error = AgentRunError(
+        agent=AgentRuntime.codex,
+        result=CommandResult(returncode=124, stdout="", stderr=leaked_stderr),
+        reason_code="AGENT_IDLE_TIMEOUT",
+    )
+
+    evidence = classify_conformance_stall(
+        history=history,
+        policy=_stall_policy(),
+        plan_path=Path("docs/awf-plans/ws_redact.md"),
+        report_path=Path("docs/awf-plans/ws_redact.conformance.json"),
+        latest_error=error,
+    )
+
+    assert evidence is not None
+    assert leaked_token not in evidence.last_output_excerpt
+    assert "<redacted>" in evidence.last_output_excerpt
+
+    payload = build_conformance_stall_failure_evidence(
+        stall=evidence,
+        head_sha=None,
+        base_sha=None,
+        commit_count=0,
+    )
+    assert leaked_token not in payload["last_output_excerpt"]
+    assert "<redacted>" in payload["last_output_excerpt"]
+
+
+@pytest.mark.unit
+def test_stall_output_excerpt_falls_back_to_redacted_exception_text() -> None:
+    record = ConformanceIterationRecord(
+        iteration=1,
+        elapsed_seconds=1,
+        report_digest=None,
+        worktree_changed=False,
+        stdout="",
+        stderr="",
+    )
+
+    excerpt = planning_mod._stall_output_excerpt(  # noqa: SLF001
+        RuntimeError("provider failed with sk-test-secret"),
+        record,
+    )
+
+    assert "<redacted>" in excerpt
+    assert "sk-test-secret" not in excerpt
+
+
+@pytest.mark.unit
+def test_evidence_strings_ignores_non_list_payloads() -> None:
+    assert _evidence_strings({"gaps": ["not", "a", "list"]}) == ()
+
+
+@pytest.mark.unit
+def test_classify_conformance_stall_returns_over_duration_for_wall_timeout_with_active_output() -> (
+    None
+):
+    history = [
+        _iter_record(
+            iteration=0,
+            elapsed_seconds=120.0,
+            report_digest="abc",
+            worktree_changed=True,
+            stdout="some output",
+        ),
+        _iter_record(
+            iteration=1,
+            elapsed_seconds=900.0,
+            report_digest="def",
+            worktree_changed=True,
+            stdout="streaming progress chunk",
+            stderr="more progress",
+            error_reason_code="AGENT_TIMEOUT",
+        ),
+    ]
+    error = AgentRunError(
+        agent=AgentRuntime.codex,
+        result=CommandResult(
+            returncode=124, stdout="streaming progress chunk", stderr="more progress"
+        ),
+        reason_code="AGENT_TIMEOUT",
+    )
+
+    evidence = classify_conformance_stall(
+        history=history,
+        policy=_stall_policy(),
+        plan_path=Path("docs/awf-plans/ws_wall_timeout.md"),
+        report_path=Path("docs/awf-plans/ws_wall_timeout.conformance.json"),
+        latest_error=error,
+    )
+
+    assert evidence is not None
+    assert evidence.kind == ConformanceStallKind.over_duration
+    assert evidence.iteration_index == 1
+    assert evidence.elapsed_seconds == pytest.approx(900.0 + 120.0)
+    assert evidence.no_output_seconds == 0.0
+    assert evidence.repeated_output_count == 0
+    assert evidence.last_report_digest == "def"
+    assert evidence.plan_path == "docs/awf-plans/ws_wall_timeout.md"
+    assert evidence.report_path == "docs/awf-plans/ws_wall_timeout.conformance.json"
+
+
+@pytest.mark.unit
+def test_classify_conformance_stall_returns_no_output_for_idle_timeout_below_policy_threshold() -> (
+    None
+):
+    # AGENT_IDLE_TIMEOUT is an explicit adapter signal (e.g. when an operator
+    # sets AWF_AGENT_IDLE_TIMEOUT_SECONDS lower than the loop's
+    # policy.no_output_seconds). The classifier must honour it regardless of the
+    # loop policy threshold, otherwise the executor re-raises the idle timeout
+    # as a generic agent failure instead of AGENT_STALLED_IN_CONFORMANCE.
+    history = [
+        _iter_record(
+            iteration=0,
+            elapsed_seconds=30.0,
+            report_digest=None,
+            worktree_changed=False,
+            stderr="idle timeout exceeded",
+            error_reason_code="AGENT_IDLE_TIMEOUT",
+        ),
+    ]
+    error = AgentRunError(
+        agent=AgentRuntime.codex,
+        result=CommandResult(returncode=124, stdout="", stderr="idle timeout exceeded"),
+        reason_code="AGENT_IDLE_TIMEOUT",
+    )
+
+    evidence = classify_conformance_stall(
+        history=history,
+        policy=_stall_policy(no_output_seconds=600),
+        plan_path=Path("docs/awf-plans/ws_no_output_below.md"),
+        report_path=Path("docs/awf-plans/ws_no_output_below.conformance.json"),
+        latest_error=error,
+    )
+
+    assert evidence is not None
+    assert evidence.kind == ConformanceStallKind.no_output
+    assert evidence.iteration_index == 0
+    assert evidence.no_output_seconds == pytest.approx(30.0)
+    assert "idle timeout" in evidence.last_output_excerpt
+
+
+@pytest.mark.unit
+def test_classify_conformance_stall_returns_no_output_when_stdout_empty_across_consecutive_iterations() -> (
+    None
+):
+    history = [
+        _iter_record(
+            iteration=0,
+            elapsed_seconds=120.0,
+            report_digest="abc",
+            worktree_changed=True,
+            stdout="some output",
+        ),
+        _iter_record(
+            iteration=1,
+            elapsed_seconds=300.0,
+            report_digest=None,
+            worktree_changed=False,
+            stdout="",
+        ),
+        _iter_record(
+            iteration=2,
+            elapsed_seconds=350.0,
+            report_digest=None,
+            worktree_changed=False,
+            stdout="   \n",
+        ),
+    ]
+
+    evidence = classify_conformance_stall(
+        history=history,
+        policy=_stall_policy(no_output_seconds=600),
+        plan_path=Path("docs/awf-plans/ws_no_output_streak.md"),
+        report_path=Path("docs/awf-plans/ws_no_output_streak.conformance.json"),
+        latest_error=None,
+    )
+
+    assert evidence is not None
+    assert evidence.kind == ConformanceStallKind.no_output
+    assert evidence.iteration_index == 2
+    assert evidence.no_output_seconds == pytest.approx(650.0)
+    assert evidence.elapsed_seconds == pytest.approx(770.0)
+    assert evidence.repeated_output_count == 0
+
+
+@pytest.mark.unit
+def test_classify_conformance_stall_returns_none_when_stdout_empty_streak_below_no_output_seconds() -> (
+    None
+):
+    history = [
+        _iter_record(
+            iteration=0,
+            elapsed_seconds=200.0,
+            report_digest=None,
+            worktree_changed=False,
+            stdout="",
+        ),
+        _iter_record(
+            iteration=1,
+            elapsed_seconds=200.0,
+            report_digest=None,
+            worktree_changed=False,
+            stdout="",
+        ),
+    ]
+
+    evidence = classify_conformance_stall(
+        history=history,
+        policy=_stall_policy(no_output_seconds=600),
+        plan_path=Path("docs/awf-plans/ws_no_output_short.md"),
+        report_path=Path("docs/awf-plans/ws_no_output_short.conformance.json"),
+        latest_error=None,
+    )
+
+    assert evidence is None
+
+
+@pytest.mark.unit
+def test_classify_conformance_stall_breaks_no_output_streak_on_stderr_progress() -> None:
+    history = [
+        _iter_record(
+            iteration=0,
+            elapsed_seconds=400.0,
+            report_digest=None,
+            worktree_changed=False,
+            stdout="",
+            stderr="progress: still working...",
+        ),
+        _iter_record(
+            iteration=1,
+            elapsed_seconds=400.0,
+            report_digest=None,
+            worktree_changed=False,
+            stdout="",
+            stderr="",
+        ),
+    ]
+
+    evidence = classify_conformance_stall(
+        history=history,
+        policy=_stall_policy(no_output_seconds=600),
+        plan_path=Path("docs/awf-plans/ws_stderr_progress.md"),
+        report_path=Path("docs/awf-plans/ws_stderr_progress.conformance.json"),
         latest_error=None,
     )
